Improved connection reconnect and command queueing
Added `cast` method to handle sending a metric when the socket has been disconnected but metrics are still being sent. Also added a queue so that metrics sent in that period are saved and sent when connection is re-established.

Basic tests for the public interface are included to prove the correctness of the public interface and network code. They are not good for the following reasons
- They use sleep to wait for connection and wait for casting, which makes them slow.
- They'll never fail because cast always return :ok. You'll need to go look at the Instrumental project where you sent the metrics to actually see if it works.
- You'll need to provide a file called `test_key` that contains API key for the Instrumental project you intend to send the metrics to. There's no code for defending against that file being missing. The error message isn't great.

On the plus side, the library does actually function better. Also, any connection testing is better than no connection testing.