Remove time return value tests
These tests asserted that the time functions would return the value of the timing. They passed originally because the Connection class wasn't started. When the genserver is started, the time functions will always return :ok because that's what the cast calls from the genserver always return.

We don't assert this functionality in our other agents, so removing for now and may revisit later.