WARNING This is a dump commit. It will not make sense.
These tests are all super duper heisen. The primary issue seems to be one of process cross-talk. In the test, you'll see that we start a server and store it's PID (`server`). Later, we expect that server to send a message to the test process containing the command it received and the PID of the server that received it (`sender`). The `server` never matches the `sender`. Not ever.

`sender` is a Supervisor and you can get a list of its children with `Supervisor.which_children(sender)` in pry. None of it's children match `sender` either. I have no idea what `sender` pid is or how to find out.

The result of this is that random tests receive random messages unless they are run by themselves.

In addition, we've found several vestigial issues in the code. Since we're not able to write any reliable tests, it seems very unwise to make many more changes to the code.

Hand testing seems to indicate that this library is happy-case ready, but real testing indicates it likely has many edges.